Fixes #20, uses getopt instead of php-cli-tools for arguments
So the GetOpt library offers a couple of advantages to the php-cli-tools
library I used before to do the argument handling for this project.

Mainly the advantages are:
1. The operand allows the users to enter at least one directory to scan
which was only possible to do with an option with the previous method.
2. Validation of option inputs, such as numbers guides the user
toward correct syntax.
3. More natural method of adding multiple arguments as an array. Before
this has to be used with comma-seperated syntax or other custom
input syntax. Now the user can just use the same parameter multiple times
such as the `--exclude` and `--file-extension` options in this package.

The command line arguments have changed a bit, so this change will be in the
v0.2.0 release. #3 will be part of the same release, so the default behaviour
only changes once.